Handle NO_EXTERNAL_FUNDING_DETAILS_FOUND
