+++ classes/Builder/Payload/OnboardingPayloadBuilder.php
@@ -22,7 +22,7 @@ namespace PrestaShop\Module\PrestashopCheckout\Builder\Payload;
 
 use PrestaShop\Module\PrestashopCheckout\Adapter\LanguageAdapter;
 use PrestaShop\Module\PrestashopCheckout\Adapter\LinkAdapter;
-use PrestaShop\Module\PrestashopCheckout\PsxDataMatrice\PsxDataMatrice;
+use PrestaShop\Module\PrestashopCheckout\PsxData\PsxDataMatrice;
 use PrestaShop\Module\PrestashopCheckout\Repository\PsAccountRepository;
 
 /**
